3 Sample Queries for SMHA database
Query 1 Identifies the first and last name, birth year, ID number, and phone number of every player on team 2, and also identifies whether that player is a skater, goalie, or both. 

Query 2 Combines all elements of addresses into a single field and provides contact information (email and phone number), the name of the location, and the number of games that are scheduled to be there. 

Query 3 Gives the name, ID, combined location, and date of the first game at each of the arenas.